Fix "Redundant call of conversion method" warnings
All of these conversions were unnecessary as it converts from the same
type to the same type.

Test: compile with KGP 2.3.0-Beta2

+++ wear/compose/compose-material3/src/main/java/androidx/wear/compose/material3/ScrollIndicator.kt
@@ -938,7 +938,7 @@ internal class TransformingLazyColumnStateAdapter(
 
             // Coerce item size to at least 1 to avoid divide by zero for zero height items
             val firstVisibleItemFraction =
-                firstVisibleItemSize.toFloat() /
+                firstVisibleItemSize /
                     (firstVisibleItem.transformedHeight + extraPadding).coerceAtLeast(1).toFloat()
             return firstVisibleItem.index.toFloat() - firstVisibleItemFraction
         } ?: 0f
